Add test suite
The state of integration of Cabal with test frameworks seems to be a
mess, although there are numerous test frameworks. Hspec appeared to be
modern and easy to use, so let's try that one. You can run "cabal test"
and it will run the tests, but it doesn't use the deeper integration
provided by Cabal.

For now there is a single test that fails at the moment.