Keep track of integration commit for the candidate
The commit from the pull request can be a different commit than the one
we want to test, because we want to test the new head, not the isolated
changes. (In the case of a fast-forward integration strategy where it is
required that all pull requests can be fast-forward merged into master,
the commit would be the same. But in general they will not be the same.)