Record time in build status
If I am going to infer actions purely from state (and not in response to
events), I need to deal with timeouts and such. Keep track of the time
when a request was sent, and if that was too long ago, make a new
request.

As I am writing this, I wonder though whether I should do this at all in
a prototype. Perhaps it is better to not deal with flaky connections and
retries for now.

The leapseconds-announced package is not on Stackage, so an exception
has been added to stack.yaml.