Define GitOperation free monad
This will replace the action type. For now there is no action to trigger
a build. Instead, I will push to a dedicated integration branch on
GitHub, and that should trigger Travis CI automatically.

This is not very flexible. For instance, it would not allow testing
every commit. But for a quick prototype this should be sufficient.