Make event handler return new state in the monad
Event handling now produces the next state in the monad, so it can
perform some action to compute the next value. This is not used yet, all
current handlers are pure. When handling an event requires for example
trying a rebase, and the sha of the rebased commit must be known to
proceed, this side effect is encapsulated.

In the tests, there now is a function that simulates doing these
actions, which keeps track of the actions that would have been
performed. This is extremely powerful for two reasons:

  * The simulator can be made to return arbitrary results for actions.
    "Mocking" of all things that are interacted with is for free.

  * Besides testing what a function does, it is possible to test what it
    does *not* do. Assertions can be added to verify that a function
    produces no side effects. (There are no such checks in the tests
    yet at this point.)