Switch to Scotty for webserver
Warp and Wai are a bit too low-level for me. I would have to check that
the request method is as expected (which is not too bad), but I would
also have to deal with chunked request bodies and the like. Scotty is a
layer on top that abstracts this away.

Initially I refrained from using Scotty because I do not like to add
extra dependencies, and I don't like things that work magically without
me understanding how they work. But the dependencies are not so bad,
Scotty generates a Wai application that that can then be served by Warp.
This Haskell web ecosystem works really nicely!