Add basic worker thread to handle webhook events
For now it only prints the webhook events. Later it should generate the
proper events for the event handler, and there should be an additional
event loop that updates the state.

I tested the bounded queue locally by introducing a one second delay
after handling an event, and then spamming the service with curl calls.
The bounded queue works beatifully. However, this is a bit hard to have
an automated test for. Perhaps extract the queues and event loops into a
component, and test that by injecting webhook events.