Add interpreter for high-level actions
The interpreter generates only high-level Git operations at the moment,
but this would be the point to do other things as well, such as
generating GitHub API calls or posting messages in a chatroom.

Additionally, add a mock interpreter for the GitOperation free monad
that simply prints what it should have done.