Get get temp directory from environment
This makes the tests portable. The end-to-end tests now run on Windows,
although removing the directory after running tests still fails with a
permission denied error. Perhaps this is because the main event loop
thread is still performing Git operations in that directory, although
event a two-second wait does not resolve the issue. Further
investigation is required.