Use a queue of Maybe Event to support termination
The solution with a Stop variant in Event is suboptimal, because it
complects event handling and the event loop: the loop needs to inspect
the events, and the event handler needs to handle a stop event.
Furthermore, a stop event allows every sender to terminate the event
loop, which is bad form a least-permissions perspective.

A better solution is to make the queue a queue of 'Maybe Event', instead
of a queue of 'Event'. 'Nothing' would be used to signal loop
termination. This still does not solve the issue of senders being able
to terminate the loop, but there is a neat solution for that: senders
should not have access to the raw queue, they should only be provided a
function that enqueues events. This is cleaner too than the previous
approach of giving senders access to the queue itself, because now
senders can only send, and not pop events from the queue.

An alternative that I considered was a separate MVar to signal loop
termination, but there is no way to wait for either an item to be
enqueued, or the mvar being signalled, so that would make the event loop
busy-wait while there are no events (which is most of the time).