Use better way of generating uuids
There was a bug in the uuid package in versions prior to 1.3.12. I have
no patch-version bound because patch-version differences should not be
API differences. Just because I am aware that a newer version exists
does not mean I should force it onto everybody (though this is a
subtle situation indeed). Stack should take care of choosing a recent
version.

Also, generating uuids using randomIO could lead to duplicates, it is
better to use V4.nextRandom, and this has been fixed.