Use a logging monad instead of raw putStrLns
This allows using logging when running the app, but discarding logs
during tests, so the output does not get polluted.

The package used for this is "monad-logger", which seemed appropriate.
It works well with the different monad classes from the "transformers"
package, and it has simple functions to run with logging to stdout or no
logging. I might want to configure the stdout message format later on,
but that is no concern at this point.