Do not rely on commit shas in end-to-end tests
Commit shas depend on the timestamp at the author and commit time. This
means that a rebase operation can produce commits with different shas,
dependin on the time it was executed.

This time dependency made the rebase test flaky, because it compared
shas of an initial rebase against those of a later rebase. If the two
rebases would happen in the same second (which happened most often), the
test would pass. But if the later rebase happened during the next
second, it would produce different shas, and the test would fail. The
solution is not to compare shas directly, but histories instead, through
their commit messages.

Because now the sha of the integration commits is not known, it has to
be retrieved from the state in the event loop, in order for a proper
"build succeeded" message to be constructed. This is currently not
implemented, so the rebase test has been disabled temporarily.