Use a semaphore for server startup synchronization
Previously a TMVar was used, and although it works on my machine,
it fails consistently on Travis, but with various combinations of
rtsopts I still cannot reproduce. The runtime claims that a thread
is waiting on a TMvar that will never be signalled, but I don't yet
understand how that could ever happen. I did observe this error
locally before, but I cannot reproduce it now.

Perhaps semaphores will sovle the issue. They are a bit cleaner anyway.