Switch from SHA + SecureMem to Cryptonite
The implementation is one line longer, and it did not reduce the string
type madness. But it does reduce the number of dependencies, as other
packages already depended on Cryptohash (the precursor of Cryptonite,
which is now just a wrapper package).

Instead of comparing base16-encoded strings, this now compares actual
bytes. Still in constant time.

The tests have deliberately not been ported yet: they still pass which
means the two implementations are compatible, so this is probably
correct.

Binary size comparison (built with --split-objs, no optimization flags):

  Implementation  Original (KiB) Stripped (KiB)
  --------------- -------------- --------------
  SHA + SecureMem           5657           3611
  Cryptonite                5615           3561

So the win is quite small, but it is a win nevertheless.

Performance comparison (no optimization flags, non-threaded runtime):

  Implementation  Avg Latency (ms) Stdev (ms) Req/sec Stdev
  --------------- ---------------- ---------- ------- -----
  SHA + SecureMem             1.00       1.30   3.75k   214
  Cryptonite                  0.96       1.49   3.97k   337

These results were obtained using wrk:

  $ wrk -c16 -d60 -t4 -s makepost.lua http://localhost:5261

Here makepost.lua generates a POST request with a valid hmac, but the
body is not valid json, so parsing it will fail. I did not put too much
effort in this benchmark, and likely I measuring the wrong thing. But at
least the evidence suggests that neither package is a lot worse than the
other.