Keep track of the ref for pull requests
GitHub provides the ref in the webhook payload. Because we cannot fetch
an individual commit given its sha, the best we can do, is to fetch the
branch that is supposed to contain the commit. This means keeping track
of the branch (ref) everywhere, so it has to be propagated from the
webhook payload into the application state, so eventually it can be used
in the "integrate" action.