Revert "Keep track of the ref for pull requests"
Using the branch name for fetching is no good, because on GitHub, the
branch may live in a different repository. It would be possible to add
that repository as a remove, and the webhook includes enough information
to do so, but there is a much cleaner way: GitHub exposes refs for every
pull request at refs/pull/:id/head. This ref can be computed from the id
of the pull request, so there is no need to keep track of the branch
name everywhere, so let's get rid of that first.

This reverts commit db8b6dc389682331bace0b048661b6e84c57eaf8.