Fetch the GitHub pull request ref before rebase
My previous approach of fetching the branch was not right, because the branch is
something in the forked repository, but the system should not know about the
forked repository and fetch from the configured one only. Fortunately, GitHub
exposes a refs/pull/:id/head for every pull request, so this can be fetched. It
also simplifies the code, as the ref name can be computed from the pull request
id, and there is no need to thread a branch name through everywhere.

The previous approach did work when I tested it, because the branch from which I
submitted the pull request was in the same repository as the target branch.

The end-to-end tests have not been updated yet.