Add a flag to build a static binary
Unfortunately it does not work for me locally, because Arch does not
distribute the required static libraries. I could build them from
source, but I was using Stack to have a simple build process in the
first place ...

If this works on Travis CI, then I could have Travis upload the binaries
to GitHub automatically. I wouldn't be able to sign them though. (Of
course I could download the binary and sign it, but that would defeat
the purpose ...)