Do not link release builds statically after all
The libraries to link against must be available at the development
machine, which can be a pain -- a pain that I was trying to avoid by
linking statically. Perhaps just compiling at the target machine is a
better solution. Or I would have to get into building in a vm or
container, which is a pain in its own right. Stack claims to make it
easy with its Docker support, but I am not sure I want to go down that
rabbit hole already.