Force line buffering on stdout and stderr
The runtime tries to be smart here and use block buffering -- which
would be great if the goal of the app was to process data as fast as
possible and write the result to stdout -- but in this case it prevents
log messages from showing up at the time they were written when running
under systemd (or just redirecting stdout to a file). So explicitly set
the buffering mode to line buffering, so log output ends up in the
journal live, with the correct timestamp.