Use 'die' instead of 'putStrLn' with 'exitFailure'
The die function prints to stderr, which is more appropriate, and
hopefully it also resolves the issue of my output not showing up in the
journal. The exitFailure function claims to flush stdout and stderr
before exiting, but I still see nothing in the journal when running
under systemd. This could be an issue with how I am using systemd
though. Still, the die function is more appropriate, so let's use it.