Add configuration option for reviewers
This does not yet enforce that only reviewers can approve proposed
changes. It only adds the field to the configuration that will allow
reviewers to be specified.

In the future, reviewers should probably include a full name and email
address, so a "Reviewed-by" line can be added to the commit message, but
for now the GitHub username is the only thing of importance.