Only accept approval stamps from reviewers
This is now based on GitHub usernames, which is probably sufficient for
now, but like with crates.io it can cause trouble if somebody changes
their username. But as long as this is configuration, and not something
that couples to your account automatically, this is a feature, not a
bug.