Always enable assertions
I care more about aborting on an inconsistent state than I care about
performance here. If the system pushes something to a repository that
it should not have pushed, that is pretty bad.

Note: the issue that assertions were no-ops before (due to the default
optimization settings, probably) was discovered through HPC: the
condition for the assertion was showing up as uncovered, and it does
show up as covered now.