Check dependency licenses on Travis CI
To avoid accidentally depending on a library that has a license
incompatible with Apache 2.0 (for instance, LGPL, because GHC links
Haskell dependencies statically), verify the list of dependency licenses
against a whitelist of allowed licenses.

Hackage is surprisingly uniform in with licenses: the vast majority of
dependency licenses is BSD3, a few are MIT, and one is public domain.
So far, so good.