Disable license check for now
The cabal-dependency-licenses package does not work with Stack, so I
could *also* install all dependencies with Cabal, but that seems
excessive.

I submitted a pull request for Stack that adds a --license option to
stack list-dependencies, so if that gets accepted, it could be used
here. For now, just disable the check.