Make writing the state file atomic
In case the application is killed when it is in the middle of writing a
new state file, it must not corrupt the state file. The last message
will be lost either way (which should be fixed with a graceful shutdown
regardless), but there is no need to lose the previous state.