Extract queue popping from main event loop
Instead of providing the event loop with the queue of events and making
it pop itself, provide it with a function that returns the next event.
This is the principle of least power: the main event loop never needs
to manipulate the event queue apart from popping from it, so by not
providing it with the queue, it can certainly never push something
onto it.