Keep track of most recent project state
The webinterface will then be able to retrieve the most recent project
state from the state var when it needs to serve the status page. In
the mean time, the event loop can just continue on its own pace in the
background. There is no need for blocking or locks or anything, perhaps
apart from the brief moment where the state var needs to be updated.