Do not show pull request number in webinterface
Less is more. I don't really want to know that number when I look at the
queue. And when I do, the number is also in the link, so just hovering
the title will bring it up. Or on mobile, long-press the link. And on
mobile there is less space anyway, so not showing the number eliminates
a potential problem there too.