Do not enable TLS by default in example config
It is nice to show that TLS is possible, and to nag people to encrypt
their connections, but it also complicates the initial setup
significantly. In particular, it changes the default from "runs but
doesn't do anything useful" to "crashes until properly configured".
That is a threshold for potential users that I don't want to have. So
default to http, such that starting the service after a fresh install
works.