Do not pass argument again in recursive call
I've had space leak problems with this in other projects. Let's
preemptively make a recursive function with a single argument, rather
than passing in the extra argument every time. By putting it in a
let-binding (or where in this case), the binding will be shared, so
we don't get thunks building up for that additional argument.