Set up correct user when cloning repo
Now it is no longer necessary to manually clone to set up a new
repository. This is a big step forward towards a zero-installation
systemd unit. Just drop in a config file, and you should be good to
go. Only currently the ssh keys and config need to be set up properly,
but I intend to address that as well too.