Make DynamicUser= *almost* work ...
After a detour of stracing the guest and debugging libnss_systemd, I
found a way to make the dynamic users usable from the guest. The only
issue now is that the ssh private key needs to be mounted, and mine is
owned by root, so it shows up as owned by root inside the container. And
I cannot chown it to the right user in advance, because with
DynamicUser= we don't know who that user is going to be.

I think a nice solution in the long-term would be to generate the key at
first run, and expose the public key through the webinterface for easy
access. But for now, unfortunately it is broken.