Delete obsolete documentation, update other docs
At first I had big plans. But building a code collaboration system would
be a full-time endeavour, and it is not what I want to focus my efforts
on. The idea now is to just be a merge bot. That is an attainable goal,
it actually works right now.

Also, it is built specifically for GitHub, nothing else.