Revert package and unit file to work on Ubuntu
I experimented a bit with running a unit on CoreOS before, from a
Squashfs image, and although I got that to work in the end, the
experience was not great, because systemd units are essentially
undebuggable, and especially with bind mounting things you get
misleading errors. Maybe some other time we can revisit that. But for
now, let's build a plain old Debian package again.