Base test expectations on git-log --graph
Previously, before the implementation of a --no-ff merge, the history
was linear, so it could be checked with "git log". But now that we do
create the merge commit, "git log" needs to flatten the history.
Apparently, it does not do this deterministically, which is why the
tests became flaky.

Resolve this by not trying to flatten the history into a linear history;
instead, check that the output of "git log --graph" is as expected. This
has the additional advantage that we can more clearly check the
structure of the graph.

For this to work the order of the history needs to be reversed, because
--graph is incompatible with --reverse, and reversing the output of
--graph means that the fork and join ascii art would be mirrored.