Replace "LGTM <sha>" syntax with "@bot merge"
Having to provide the sha of the commit to merge, while good to double
check that you merge what you intended to merge, is not user-friendly,
especially not on GitHub which lists comits chronologically rather than
in graph order, and selecting and copying the sha from the UI requires
some fiddling.

On the other hand, we do not want to trigger on arbitrary comments, so
add a configurable prefix. In most cases we would set this prefix to the
username of the bot. E.g. "@hoffbot ". Then "@hoffbot merge" would
trigger a merge.

This commit adds the configuation and updates the check, but it does not
yet fix the tests or example config.