Enable -Wincomplete-uni-patterns
When I originally wrote Hoff I was not aware of this warning. It
prevents bugs, without it you can get match failures at runtime. It is
not included in -Wall though, it must be enabled explicitly.

Enable it in the library and binary but not in the tests; in the tests
it is sometimes useful to do let-bind a specific constructor. That keeps
the tests brief, and in tests it acts as an assertion and and binding in
one. A runtime error in the test suite is not an issue, that will just
make the tests fail.