Drop trailing space from trigger config
Having a trailing space in configuration is a bit of a footgun. In json
it is at least visible, though easy to miss, but imagine that that line
was generated by something (say, yaml without quoted strings *shudder*),
then it may not be visible at all. And either way, it is unexpected.