Add wrapper type for GitHub usernames
As a preparation for doing the auth check through the API, instead of
having a list of usernames in the configuration, we can improve the type
safety of working with usernames with a wrapper type.

This does not change the schema, because the {From,To}JSON instance just
renders the string itself.