Do not hard-code interpreters in runLogicEventLoop
In the end-to-end tests, we want to run with the real IO interpreter for
Git actions to test the Git integration, but we don't want to make API
calls to the GitHub API. By passing in the interpreters as arguments, we
can pass in the real one to handle Git, and a fake one to handle GitHub.

This does not yet update the tests, only the binaries.