Simplify Spec by using foldFree and tuple monad
No need to match on the Free, we can use foldFree. But in order to use
foldFree, we need a function that interprets to a monad. Fortunately,
a tuple of list of ActionFlat with something, is a monad. And we can use
that, instead of having to write our own prepend function.

What is a bit of a shame is that the order of the tuple elements is
reversed. Fixed now with a swap, to ensure a smaller diff. We can get
rid of it later.