Fuse fixed-point iteration and event handling
After handling an event, we need to call "proceed" until the state
reaches a fixed point, to really finish handling that event. There are
some intermediate states that we only expect to observe when the state
is not a fixed point of "proceed".

In the tests, when we handle an event, we really want to test the full
event handling, up to the fixed point, not just the initial step that
kicks off the process.

Yet, some of the tests focus on handling the later part, and that is
fine, I don't want to have to add events there to massage the state into
the right initial state; we can just set it up like that and check that
"advance" works as expected.

To make things clearer, make "handleEvent" handle the event and also
proceed until a fixed point. There is now "handleEventInternal" that
only performs the initial actions. Update the tests accordingly.