Fix queue position reporting
This is a bit more complex than I would like it to be. This happens
because a pull requests has different fields for different statuses, but
not all combinations of these are valid. For example, BuildNotStarted is
not a valid state for an integration candidate; the integration
candidate should always be BuildPending. If it is not, then it should
not be the candidate. At least, in a fixed point it should not be.

We could maybe solve this by having a nested ADT, with e.g. the build
status only tracked for PRs that have been integrated. But let's do that
another time.