Do not delete remote branch after merging
Automatic branch deletion led to problems when the branch was still in
use as a base branch of another pull request on GitHub.
In that case, GitHub won't allow deleting a branch through its web
interface, but it can still be removed by pushing a branch deletion.
That branch deletion causes GitHub to close all pull requests that used
the branch as base.

Therefore, we are better off not deleting remote branches in hoff and
instead relying on GitHubs automatic branch deletion, which will handle
this case correctly.