Ignore pull request commit change if we caused it
This fixes a bug where:

 * The test build succeeded, we want to push the integration candidate
   to master.

 * We first force-push to the pull request branch, so GitHub marks the
   pull request as merged instead of closed.

 * We push to master, but the push is rejected, because something else
   was pushed to master in the mean time.

 * No problem, that clears the integration candidate and the loop
   restarts.

 * But, in the mean time, GitHub delivers a Synchronize webhook, that
   says that the commit hash for a pull request has changed. (Which it
   did, we pushed it ourselves.) And then we treat the pull request as
   unapproved.

Usually we would want a new commit hash to indeed clear the approval
status, but not when we caused the push ourselves. In that case, ignore
the push. This way we also remember the initial commit hash that we were
trying to rebase, and we try rebasing the original commit, instead of
the rebased commit.