Improve rebase failure message
Sometimes, rebasing only fails due to `--autosquash`, but a non-autosquash rebase is still possible.
In that case, hoff complains that rebasing is not possible, even though it appears to work fine when rebasing manually.
By including the shape of the rebase command in the error message,
the user is more likely to discover the problem why the rebase failed.