Add readonly interpreters for Git and GitHub
I want to run Hoff with a real config in local development. But I don't
want it to start merging things, or leaving comments. What would be
useful for this is to have a "dry run" or "read only" mode, that does
not perform side effects like pushing or leaving comments, but which
does run everything else.

As a first step, add interpreters for this dry run. They override the
dangerous operations with log-only ones, and forward to the real
interpreters for the safe operations.