Add basic optparse parser for main binary
I want to add a --read-only flag, and bringing in optparse-applicative
is easier than hand-rolling the mix of positional arguments and flags.
We use optparse-applicative in other projects, it is nice to use,
stable, and it I don't think it has ever cause us problems. It
definitely solves more problems than it causes, so the dependency is
worth it.