Echo command line options at startup
This can save a lot of debugging time, when you figure out that some
option was not interpreted as expected, or when you were reading the
wrong config file. Right now, I want it because I want to be sure that
it is going to run in read-only mode, so let's first confirm that the
options parser works.