Handle 404 response from pull request endpoint
If the pull request does not exist (for example, because the thing with
that id is actually an issue), treat it as non-existent.

This should not normally happen in production; it happened during
testing because I manually edited the state to include some issues. In
production these would not end up in the state in the first place.

I don't know if there is a different scenario that can trigger this
though, so let's handle it for robustness.