Retrieve missing pull requests on synchronize
For this, add a new operation to get the details of a pull request. We
can recycle the old operation to get its state, which is now no longer
used.

Also edit the test interpreter, so we can have control over the results
it returns, without having to pass all results all the time, by having a
record with a default, so we can override results as desired.