Add warning that repository is public to templates
Hoff controls our private repositories, so if we discover a bug or
discuss a feature, it often involves private details. Also, the logs
contain comments and pull request titles from private repositories.
We don't want to expose those in the public issue tracker, add a warning
for that to the pull request and issue template.