Change the wording of the second pull request
The previous wording, "waiting for rebase at the front of the queue",
can be confusing. It is not clear what queue is referred to here. Since
it is unclear whether the queue includes the pull request that is
actively being merged, being "at the front of the queue" can easily be
interpreted as being processed immediately.

This is also more consistent with the `n` pull requests.