Use nix for stack build
The recent nixification of the release build introduced some
incompatibilities between the dependency versions used for the release
build (via nixpkgs) and the development build (downloaded via stack).

By using stack's nix-mode we can use the same set of dependencies for
all builds (and get the binary cache of Haskell dependencies for free).