refuse to merge an empty rebase
... in order to avoid Hoff hanging indefinitely while waiting for the CI
result.

Here is the original description of the issue by @fatho:

In rare occasions it can happen that rebasing a branch on top of master leaves
it empty. This is the case when all the changes from the branch have already
been applied to master in the meantime. Git will then omit the empty commits,
and have the branch point to the latest commit on master directly.

But since that way, no new commit is generated, no CI job is spawned, and thus
Hoff never gets the webhook for the CI result, causing it to hang indefinitely.

When encountering an empty branch after rebasing, Hoff should post a comment
and abandon that PR.

For more details, please see:

https://github.com/channable/hoff/issues/75