Implement the parallel-builds Bayesian strategy
It does not do as well as I had hoped. But after thinking about this
problem for a few days, I have a feeling that we can do better anyway.
Right now we try to reduce the time spent waiting for build build in one
greedy step ... but if you know that you have more build slots
available, wouldn't it be better to split that step up into two that
build on top of each other?