infer the integrationCandidate (w/ failing tests)
Infer the `integrationCandidate` instead of storing it explicitly on the
project state.

This may make it easier to implement the merge train as we will need to handle
multiple integration candidates.

This specific change tries to maintain backward compatibility where possible,
this backwards compatibility will gradually be cleaned up in further commits.

The code typechecks but the tests fail for now.