add and use integratedPullRequests
where applicable to replace candidatePullRequests.

In the code "candidate" can mean two different (but related things):

* a candidate for integration, as returned by the function
  candidatePullRequests;

* something that has been integrated and is waiting for build results to become
  master integrationCandidate field of ProjectState;

If I understand correctly, the latter is perhaps better described as a PR that
is "integrated-but-waiting-for-build-results" or an integratedCandidate.
The names have not yet been updated to reflect this.

cf: https://github.com/channable/hoff/issues/77#issuecomment-1184524328

This fixes some failing tests, but there are still quite a few to go...