Adjust prior for p_is_good downward
Observation: The Bayesian strategy works better with a lower prior
is_good probability. 0.9 was higher than the 0.85 true value, 0.8 is
lower. I wonder, will it perform best at the "true" value? Or does it
get rewarded for being more conservative?

At the true value, it performs in between, more conservative than the
true value is actually better.

But around 0.7, near the critical point, and more so with more
parallelism, this starts to be worse than the true value.

Perhaps, I should do a conservative estimate of the true value based on
observed data so far, that would kind of get rid of one more parameter.