Restructure Bayesian updates
We now recompute the probabilities every time the set of failed builds
changes. This can cause probabilities per PR to go up, as well as go
down. Previously they could only go down. Say that e.g. a build with
pull request A and B fails. Then A and B are both pretty suspect. But
then later we confirm that A is certainly bad. Now the fact that {A,B}
failed is no longer so surprising, and we shouldn't blame B so much for
it.

I do wonder though, does the order in which the evidence comes in
matter? It would be nice if it didn't, but I suspect it does ...