Implement alternative Bayesian update
I am very confused now. The order in which we apply evidence matters ...
a lot! Is that a good thing? I'm not exactly sure ...

One approach would be to just run the strategy with the different
updates, and see which one performs best. There are many different
orders to try, it’s something we could figure out empirically.

But on the other hand, I am not satisfied. I don't understand very
deeply where the difference comes and what it means, only that it's
there. Why would one order be better than another?