Add a function to compute subset probabilities
The function is surprisingly short and elegant. I like this superset
approach!

So with this, I could already make a strategy that maximizes the number
of pull requests *merged*. But that's not necessarily the thing that
shrinks the backlog the most, as we saw before. So I would also like to
compute the expected entropy reduction. And I have a feeling that there
must also be a simple and elegant formulation for that, but every time I
start to think about how to do it, it becomes unwieldy ...