Add function to compute expected entropy delta
And it appears to work in my manual example! With the four pull
requests, after concluding that 1 is good, including or excluding 1 does
not make any difference for the subsets (as expected, you cannot learn
more by including it). Then after we build two together but leave a
third one out, and it fails, there are the two good options:

 * Build the third one plus one of the two from the previous batch. If
   it works we merge two, if it fails we merge zero.

 * Build just one of the two from the previous batch. If it works, we
   merge one and conclude one is bad, if it fails we conclude one is
   bad. And indeed the expected entropy decrease here is a slightly
   larger.