Implement entropy-minimizing sequential strategy
And as I implement this, I think I also know what extension I really
want to take wait time into account. The remaining entropy is roughly
the number of builds we still need. So we want to minimize the remaining
entropy times the number of PRs minus the PRs we expect to merge. That
is the remaining total wait time.