Avoid using --single-branch in tests
... by making sure that the local repository is cloned at the right time.

The --single-branch argument does a bit more than just ordering git to clone
only the given branch, but actually changes the way the repository is tracked
so that only the given (master) branch is tracked.  This makes it so that tests
behave differently from the actual running app, notably when dealing with
parallel and speculative builds as in #77.

This change makes it so that tests and running code are more similar.