Fix unintegration of other PRs upon PR closure
Before an early PR closure (or new commit) had the power to unintegrate PRs in
the train, regardless if this PR was integrated or not!  The unintegration of
later PRs should only happen when the PR being closed has been integrated and
thus is the base for these following PRs.