Rewrite merge command parser using megaparsec
This maintains some of the leniency of the old parser by allowing
case-insensitive matches on anything other than the environment names
and by allowing duplicate whitespace, but at the same time it also
avoids mistakes by parsing greedily without backtracking in most cases
and by disallowing commands to be followed by anything other than
punctuation and whitespace. That means commands can still be put at the
end of a (non-wrapped) sentence, but not in the middle of a paragraph.

The new tests and the changes to the existing tests allow and reject the
specific cases mentioned in #212.