Add a retry command for PRs with failed checks
This adds a new `retry[ on friday]` command that allows approved PRs
with failed checks to resubmitted as integration candidates, adding them
to the end of the merge train. This is functionally equivalent to the
current workaround of closing the PR, reopening it, and the then issuing
t he approval command again.

An upside to having this be a separate command rather than an extra
behavior for the merge command is that you can't accidentally change the
approval type (e.g. from `merge and deploy` to plain `merge`) and it's a
very explicit action you need to take. The error messages have also been
updated to suggest the retry command instead of closing and reopening
the PR and issuing a new merge command.

`Pr.approval` now also stores who retried the PR if it has been retried.
This keeps the original approver in tact and the retrier is also shown
on the dashboard.

This resolves #214.