Migrate default.nix to Cabal and overlays
This gets rid of Stack and replaces it with Cabal. The
`nixpkgs-pinned.nix` file was also changed to the newer overlay style.
Note how the `git` argument in `hoff.nix` had to be changed to
`pkgs.hoff` since the derivation is now built using
`pkgs.haskellPackages.callPackage` instead of `pkgs.callPackage`.